Tested on Linux ARM

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: cnguyen- <cnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 23:46:33 by cnguyen-          #+#    #+#             */
-/*   Updated: 2024/05/29 19:08:19 by cnguyen-         ###   ########.fr       */
+/*   Updated: 2024/05/30 19:47:52 by cnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ typedef struct s_specs
 	int		blank; // true 1 or false 0
 	int		plus; // true 1 or false 0 
 	int		star_width; // -1 none, 0 next arg
-	int		width; // negative overflow, 0 none, 1, 2, 3, 4 ... value
+	long	width; // negative overflow, 0 none, 1, 2, 3, 4 ... value
 	int		star_precis; // -1 none, 0 next arg
 	long	precis; // - 1 none, 0, 1, 2, 3... value
 	char	specif; // c, s, p, d, i, u, x, X, %, o, unknown
